commit, remove logging, minor renaming
