+++ ui/.env
@@ -3,7 +3,7 @@
 # These settings are loaded at build time.  So rebuild after any changes.
 
 # The base URL for all requests to the REST Server.
-REACT_APP_REST_SERVER_URL=https://localhost:3200
+REACT_APP_REST_SERVER_URL=http://10.211.55.5:3200
 
 # The timeout for OAuth redirect in minutes.  After this time, the redirect
 # is ignored, and the user is sent to the networks list screen (ms).
